Move function for validating json strings to utils.py
